+++ src/compliance_tests/TestIso_7_5_4.cpp
@@ -113,7 +113,7 @@ class TestIso_7_5_4 : public test::TestBase
             drv_bit_frm_2->InsertPasErrFrm(7, BitKind::Data);
             mon_bit_frm_2->InsertPasErrFrm(7, BitKind::Data);
 
-            for (int i = 0; i < 8; i++)
+            for (size_t i = 0; i < 8; i++)
             {
                 mon_bit_frm->AppendBitFrame(mon_bit_frm_2.get());
                 drv_bit_frm->AppendBitFrame(drv_bit_frm_2.get());
