+++ src/compliance_tests/TestIso_7_6_12.cpp
@@ -75,7 +75,7 @@ class TestIso_7_6_12 : public test::TestBase
         void ConfigureTest()
         {
             FillTestVariants(VariantMatchType::CommonAndFd);
-            for (int i = 0; i < 2; i++)
+            for (size_t i = 0; i < 2; i++)
             {
                 AddElemTest(TestVariant::Common, ElemTest(i + 1, FrameKind::Can20));
                 AddElemTest(TestVariant::CanFdEna, ElemTest(i + 1, FrameKind::CanFd));
@@ -111,13 +111,13 @@ class TestIso_7_6_12 : public test::TestBase
             drv_bit_frm->InsertActErrFrm(7, BitKind::Data);
 
             /* Force n-th bit of Error Delimiter to dominant! */
-            int bit_to_corrupt;
+            size_t bit_to_corrupt;
             if (elem_test.index_ == 1)
                 bit_to_corrupt = 2;
             else
                 bit_to_corrupt = 7;
 
-            TestMessage("Forcing Error delimiter bit %d to Dominant", bit_to_corrupt);
+            TestMessage("Forcing Error delimiter bit %zu to Dominant", bit_to_corrupt);
             Bit *bit = drv_bit_frm->GetBitOf(bit_to_corrupt - 1, BitKind::ErrDelim);
             bit->val_ = BitVal::Dominant;
